Remove default implementation of `getDebugInstance`
We have to make a judgement call per occurance to determine what we
should return here.

This whole mess could use some simplification, but we probably just
won't end up keeping this around for much longer once/if we remove
the action modifier, which is currently the only kind of modifier
we have internally (since `{{on}}` is implemented in glimmer-vm and
essentially duplicates this whole "internal" set up once more time
over there).