Introduce `keywords` option for `precompile`
Recently in #1557 we started emitting build time error for strict
mode templates containing undefined references. Previously these
were not handled until runtime and so we can only emit runtime
errors.

However, the previous setup allowed for the host environment to
implement additional keywords – these are resolved at runtime with
the `lookupBuildInHelper` and `lookupBuiltInModifier` hooks, and
the error is only emitted if these hooks returned `null`.

To allow for this possibility, `precompile` now accept an option
for additional strict mode keywords that the host environment is
expected to provide.