CI Cleanup
I searched through our entire dependency graph and couldn't find anything consuming DISABLE_SOURCE_MAPS or BROCCOLI_ENV. I think they are entirely vestigial.

DEBUG_RENDER_TREE was being set in an irrelevant place, because it doesn't take effect during the build, it takes effect during the test run. But when I tried to enable it there, I found that it never has any effect anyway. The issue is that DEBUG_RENDER_TREE is always enabled in dev mode, and even when the test suite is built in prod mode it still loads ember-template-compiler.js, which always includes dev-mode code that forces DEBUG_RENDER_TREE to on. So there's no way to run the test suite with DEBUG_RENDER_TREE off, so there's no need to have an extra test that tries to turn it on.