Cleanup warning noise from eslint
We inherit several default rules at the "warn" level. IMO it's useless to leave rules at "warn", because people just ignore them and live with an increasing amount of noise.

As evidence I cite the 716 warnings that one must currently scroll through to find an actual lint failure.

This change turns unused-vars into errors, while enabling underscore-based exceptions that follow typescript's own defaults for its internal unused-vars rule.

And it disables no-explicit-any and no-non-null-assertion. These features are in the language for a reason. Idiomatic javascript that gains types after-the-fact needs these features, which is why we need them all over the place. If somebody else thinks these rules are good, I'm all in favor of enabling them **as errors** so they actually get followed, but then you've gotta do the project to refactor all the types to not need them so much.