Don't emit empty field initializer for `set`
This is arguably just cleaner style, but the reason I care in particular is that there is a bug in babel such that under certain settings this emits:

```js
class Registry {
  // ...
  set
  constructor() {
    // ...
  }
}
```

which is a syntax error because it looks like your constructor is a setter.

With `declare`, typescript completely removes the pointless runtime field initializer.