@glimmer/component is in both deps and peers
It doesn't really make sense to have the same dependency in both peerDependencies and dependencies. I think this was introduced as a mistake when the various `@glimmer` packages were moved to dependencies for typescript support.

This PR keeps the peerDependency, removes the dependency, and introduces a devDependency so we can satisfy our own peer requirement in development.