Bugfix: restore publishing of internals entrypoints
This is a followup bugfix to https://github.com/emberjs/ember.js/pull/20675.

That PR inadvertently stopped publishing entrypoints for certain internal packages, when consuming ember as separate modules. These packages are not public API, but they're used, so we need to keep them.

A symptom of what breaks here is ember-data importing `@ember/-internals/metal`, when building with Embroider's strictEmberSource option.