Break eager cycle for "deprecate" function
There is a module cycle involving `@ember/debug/index` and `@ember/debug/lib/deprecate`, and it's somewhat intentional since both are sharing mutable state with the other that's intended to allow users to replace the implementations.

The cycle wouldn't be a big deal except for the fact that index **eagerly** consumes the `deprecate` function from `lib/deprecate`, which means that if you get unlucky in the evaluation order it just explodes.

There's no need to eagerly consume it. This PR makes the consumption lazy.

In general `export function` is resilient to cycles in a way that `export let` is not. There are a bunch more places in `@ember/debug` that are using `export let` when they should probably be using `export function` instead, especially because these function are supposed to *always* be present (even in prod they are no-ops, not missing).