add v2 addon metadata
This adds v2-addon-style metadata to the package.json at publication. Our build already produces everything needed to be a valid v2 addon, other than this metadata.

This PR doesn't actually flip the final switch though (by adding `version: 2` to the metdata). So it's safe and nonbreaking. This just gets us one step closer.

And it's helpful because embroider users will be able to opt in to treating ember-source as v2 even before we force it to be v2 for everyone.