Another attempt to fix barrel.ts
This time, use an ambient `declare namespace` to declare the
`Ember.RSVP` types, which will avoid the ill-fated attempt to transpile
it at runtime using the babel typescript transform. The previous
implementation already required us to dynamically install the `RSVP`
property on the `Ember` object because the previous transpiler
infrastructure didn't reliably handle this situation, so this is
probably an improvement either way.