Don't try to fix dynamic modifiers yet
This case is not fixed by the current PR:

```
<p {{some.modifier}}>
```

It probably *should* result in `some.modifier` appearing as the `name`
in the debug modifier tree, but the codepaths improved in the current PR
don't fix that case. For this PR, we'll use the debug name provided by
the modifier manager, which is at least not a regression.

A future PR should systematically fix all of the cases that end up in
the debug tree so that they *always* show "what the user typed",
regardless of how it got converted into running code.