[CLEANUP] Remove old code that supported old ember-test-helpers
The actual behavior change was marked with an intimate deprecation
and has aged out. The actual versions of ember-test-helpers affected
seems to be around <= 3.2.0, which should be fully absorbed by now.

The other changes to `OutletState` doesn't actually change behavior,
those fields were mostly just kept around assuming we are keeping
this code frozen and to be deleted soon, so they documented what
those were on the off chance that someone came into look. Now that
it looks like this code may be here to stay for a bit longer, it's
worth cleaning it up to make things less confusing for core devs.

Follow-up to #20570