Attempt to overhaul tsconfig/eslint
1. Migrate to ESLint 9 + new config format
2. Migrate to a single tsconfig file for the entire repo

In general, this results in more consistency across the repo, including
in one-off packages and tests, but it still gives us flexibility to
create special rules if necessary.

As part of this process, a new `repo-metadata` package was created that
allows other infrastructure packages to get workspace metadata. This
package has a static JSON file with the information about each package
in the workspace. This was important in the current PR because it allows
our eslint config to be sensitive to information in individual packages
without slowing down the boot time of eslint.

Fully migrating to ESLint 9 required:

- Making the whole repo clean with `eslint .`
- Making `eslint .` in individual packages work (and perhaps move the
  custom configuration to configs in each package
- Moving the guts of the code into `@glimmer-workspace/eslint-plugin`.
  This should be **much** more straightforward with ESLint 9.
- Porting all of the ESLint rules from the current setup to the new
  config, or deciding to ignore them.
- Updating our JSON linting

Rather than using a workspace package for benchmarks and trying to
get `vite build` to emulate the expected behavior, we `pack` the
dependencies and refer to them directly in the `package.json`.

This is a little bit more involved than it might seem like it should be
because the benchmark has to run on the `main` branch, so it has to be
very general and can't use any infrastructure in the workspace itself
that could affect the builds.

On the bright side, that constraint pushes us towards treating the
benchmark environment as a normal consumer of the normal Glimmer
packages, which is good.

The only tiny exception to that rule is that the build script builds the
`@glimmer-workspace/benchmark-env` package, even though it's not a
public package. This means that the code in
`@glimmer-workspace/benchmark-env` needs to make sure it can be built.

The current setup allows us to use the `main` branch's
`@glimmer-workspace/benchmark-env` package with the experiment's
`krausest` package, so it seems reasonable to build the benchmark env
and use it as a traditional package rather than trying to get two
workspaces to somehow mesh together (which is what got us into trouble
in the first place).

Note: Getting this working right now requires a little bit of hackery
when setting up the current `control` (i.e. `main`) branch, because
`main` isn't set up to build the `@glimmer-workspace/benchmark-env`.
Once this PR is merged, we should be able to remove those hacks and rely
on vanilla package builds for the entire setup going forward.

Other changes:

- Document recommendations
- Use LFS
- Explain recommended extensions for vscode
- Remove unused dependencies

NOTE: This PR comes after multiple attempts to set up a `.meta-updater`
config, but I realized that it would be much easier after a long-needed
overhaul to the tsconfig and eslint configs.