Improve meta-updater and turbo workflows
- `meta-updater` now automates the process of keeping packages aligned
  with `publint` requirements based on whether they're published.
- add back a `-f` that was accidentally removed from the bench setup
- Created top-level `repo:*` scripts for `turbo` commands so that
  there's no confusion between the repo-wide commands and the same-named
  commands in packages.
- Added `@bandaid` highlighting rule to capture situations where code
  exists as a temporary workaround, and allows you to document the
  reason for the workaround so it can be identified and cleaned up
  later.
- Update GitHub actions to align with the new `turbo` and repo-wide
  workflows.
- Improve docs in `workspace-management.md` to describe the new
  turbo workflows and repo-wide commands.

The tests were failing because (apparently) vite wasn't analyzing the
`index.html` file properly. This resulted in multiple copies of the
validator code getting included in the tests (via chunks and via the
`.ts` module itself).

To make matters worse, the validator code included in the chunk was
missing the code that detects duplicates, so the only failure symptom
was an error reporting that the `COMPUTE` symbol wasn't present on an
instance of `Tag`.

Being explicit about entry points fixed the problem and made the tests
pass. I don't know why this is necessary (surely vite should undertand
its own import.meta.glob), but it's a relatively small change, and I
think we can live with it for the moment.