Update to ember-cli v6.2
ember-cli is used in the tests, and some tests need the newer features to work as expected.
We also need a version that uses the `babel-remove-types` package that includes the fixes for .gts detyping to work.

Because of this update, some slight adjustments are needed to some pod related tests. The `usePods` option of ember-cli-blueprint-test-helpers' `setupPodConfig` util isn't compatible with newer ember-cli versions. As a workaround we use the `--pod` argument when generating files.