Fix source maps in published packages
I noticed this problem from this warning in my build:

```
Warning: ignoring input sourcemap for ../rewritten-packages/@embroider/synthesized-vendor/vendor/ember/ember-testing.js because
ENOENT: no such file or directory, open '.../node_modules/.embroider/rewritten-packages/@embroider/synthesized-vendor/vendor/ember/ember-template-compiler.js.map'
```

Indeed, it seems like as of 5.10.0-beta.1, we stopped publishing
source maps:

https://unpkg.com/browse/ember-source@5.10.0-beta.1/dist/

...despite having every intention to do so still.

Last good version:

https://unpkg.com/browse/ember-source@5.10.0-alpha.4/dist/

It took me way to long to spot this, but the problem is in the
extension; in the broccoli build, they had a `$name.map` filename,
whereas in the new build (#20675) it's `$name.js.map`.

Adjusted the entries in package.json to match; also removed the
types publish script that was accidentally included in the tarball.

* * *

As for my original quest to fix the warning, it turns out this is
not sufficient. The file is placed there by @embroider/compat these
days:

https://github.com/embroider-build/embroider/pull/2164

It'll have to be updated to also emit the source map file (if it is
present, since this has been broken for a while).

But this is a prerequisite to fixing that.