Fix source slicing for whitespace-stripping comments
Re-apply #1430 with a more robust implementation and tests

Also turns the warning into a `localAssert()` so it gets stripped
before reaching downstream consumers.

This relates to #1431 in that this removes the warning, so to the
extent that #1431 was motivated by sliencing the warning in Ember
builds, it avoids the issue. But I haven't investigated the root
cause of that issue and whether it is indicative of an actual bug
somewhere deeper.

Fixes emberjs/ember.js#19392

Co-authored-by: Lee Nave <lee.nave@ovation.io>