set hoist-workspace-packages=false
This is part of the effort to harmonize the ember and glimmer monorepos. 

We upgraded ember to use pnpm 10 to match this repo, but discovered that starting at pnpm 9, pnpm defaults to hoisting all workspace packages such that they are erroneously resolvable by *all* dependencies. 

This setting is more strict and correct and avoids specific confusing problems in ember's build.