Fix LOG_VERSIONS
Ember's debug logging of its own version was accidentally broken since 5.10 when we reformed the build.

The problem was:
 - we switched to typescript's more spec-compliant class field semantics, where fields without `declare` are in fact present as real class fields with an undefined initializer.
 - our test suite was accidentally getting processed by vite's default (and pretty broken!) esbuild TS loader, which uses the loose & wrong field semantics.
 - `logVersions` in `class Libraries` is written in a picky way that breaks if you get the spec-compliant field semantics.

This PR disables vite's built-in esbuild, which revealed the bug.

The change to logVersions is the only real user-facing bugfix. There were also some similar bugs in the test suite itself which are fixed here.

(All of this was precipated by me working on stable decorators, which *also* requires disabling vite's hidden-and-terrible TS support.)